feat(payments): WIP Stripe integration

+++ shared/src/main/kotlin/me/rasztabiga/thesis/shared/domain/command/command/CreateOrderPaymentCommand.kt
@@ -15,6 +15,6 @@ data class CreateOrderPaymentCommand(
     data class OrderItem(
         val name: String,
         val quantity: Int,
-        val unitPrice: BigDecimal
+        val unitPrice: Double
     )
 }
